requirements.txt
flask
gunicorn app:app
psycopg2-binary
python-dotenv
requests

